refactored about page routing
