+++ webapp/src/pages/AboutBlockchain.tsx
@@ -1,0 +1,17 @@
+import React from "react";
+import {Group }from "@mantine/core";
+import { ActionsGrid } from "../components/actionsGrid";
+  
+
+const AboutBlockchain = () => {
+  return (
+    <>
+        <div style={{marginTop: 40}}></div>
+        <ActionsGrid/>
+
+      
+    </>
+  );
+};
+
+export default AboutBlockchain;
