+++ webapp/src/pages/App.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { MantineProvider } from "@mantine/core";
+import { MantineProvider, Text } from "@mantine/core";
 import { HeaderSimple } from "../components/header";
 import { Routes, Route } from "react-router-dom";
 import DAO from "./DAO";
@@ -10,6 +10,8 @@ import Education from "./Education";
 import StoreFront from "./StoreFront";
 import theme from "./theme";
 import Deploy from "./deploy";
+import AboutKUBI from "./AboutKUBI";
+import AboutBlockchain from "./AboutBlockchain";
 
 const link = [
   { link: "/", label: "Home" },
@@ -21,7 +23,15 @@ const link = [
 ];
 function App() {
   return (
-    <MantineProvider withGlobalStyles withNormalizeCSS theme={theme}>
+    <MantineProvider withGlobalStyles withNormalizeCSS theme={{
+      fontFamily: 'Verdana, sans-serif',
+      fontFamilyMonospace: 'Monaco, Courier, monospace',
+      headings: { fontFamily: 'Greycliff CF, sans-serif' },
+      colorScheme: 'dark'
+    }}>
+      <MantineProvider inherit>
+        
+      
       <HeaderSimple links={link}></HeaderSimple>
       <Routes>
         <Route path="/" element={<Home />} />
@@ -31,7 +41,11 @@ function App() {
         <Route path="/DAO/deploy" element={<Deploy />} />
         <Route path="/education" element={<Education />} />
         <Route path="/storefront" element={<StoreFront />} />
+        <Route path="/aboutBlockchain" element={<AboutBlockchain />} />
+        <Route path="/aboutKUBI" element={<AboutKUBI />} />
+        
       </Routes>
+      </MantineProvider>
     </MantineProvider>
   );
 }
