+++ webapp/src/pages/Education.tsx
@@ -6,8 +6,8 @@ import { ActionsGrid } from "../components/actionsGrid";
 const Education = () => {
   return (
     <>
-        <div style={{marginTop: 40}}></div>
-        <ActionsGrid/>
+        <div style={{marginTop: 40}}> This page is outdated, we are working to fix routing before deleting this</div>
+       
 
       
     </>
